fix: solve permissions' translation
