+++ identity/client/src/utility/localization/en/index.ts
@@ -1,5 +1,5 @@
 import components from "./components.json";
-import permissions from "./components.json";
+import permissions from "./permissions.json";
 
 export default {
   components,
